% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosswalk.R
\name{brfss_crosswalk}
\alias{brfss_crosswalk}
\title{Load the BRFSS crosswalk into memory.}
\usage{
brfss_crosswalk(path = NULL, years = NULL)
}
\arguments{
\item{path}{Directory containing the CSVs. Defaults to NULL, which automatically
finds the bundled crosswalk inside the installed brfssTools package.}

\item{years}{Optional integer vector of years to load (e.g., c(2018, 2019)).}
}
\value{
A list of tibbles representing the crosswalk schema.
}
\description{
Load the BRFSS crosswalk into memory.
}
