% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosswalk.R
\name{brfss_lookup}
\alias{brfss_lookup}
\title{Pull the full crosswalk for one or more concepts}
\usage{
brfss_lookup(cw, concept_id)
}
\arguments{
\item{cw}{A loaded crosswalk object.}

\item{concept_id}{Character vector of concept IDs.}
}
\value{
A tibble mapping the concept(s) to all relevant years and sources.
}
\description{
Pull the full crosswalk for one or more concepts
}
