% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosswalk_io.R
\name{cw_verify}
\alias{cw_verify}
\title{Mark a (concept, year) cell as verified}
\usage{
cw_verify(bundle, concept_id, year)
}
\arguments{
\item{bundle}{Bundle from \code{cw_load()}.}

\item{concept_id}{Concept name.}

\item{year}{Integer year.}
}
\value{
Updated bundle.
}
\description{
Mark a (concept, year) cell as verified
}
