% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{.brfss_package_extdata}
\alias{.brfss_package_extdata}
\title{Locate the installed package's extdata directory}
\usage{
.brfss_package_extdata()
}
\description{
Locate the installed package's extdata directory
}
\keyword{internal}
