% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosswalk_io.R
\name{.cw_find}
\alias{.cw_find}
\title{Find the row index in the crosswalk for (year, state_var, cdc_var)}
\usage{
.cw_find(bundle, year, state_var = NULL, cdc_var = NULL)
}
\description{
Find the row index in the crosswalk for (year, state_var, cdc_var)
}
\keyword{internal}
