% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosswalk_io.R
\name{.cw_v2_columns}
\alias{.cw_v2_columns}
\title{Canonical column ordering for v0.2.0 crosswalk}
\usage{
.cw_v2_columns()
}
\description{
Canonical column ordering for v0.2.0 crosswalk
}
\keyword{internal}
