% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosswalk.R
\name{brfss_harmonize_vec}
\alias{brfss_harmonize_vec}
\title{Apply a recode rule to a raw-value vector.}
\usage{
brfss_harmonize_vec(
  x,
  recode_type,
  recode_rule = NA_character_,
  is_missing_codes = character()
)
}
\arguments{
\item{x}{Raw value vector.}

\item{recode_type}{"identity", "inline", or "function".}

\item{recode_rule}{The string rule or function name.}

\item{is_missing_codes}{Character vector of raw codes to force to NA.}
}
\value{
A character vector of harmonized values.
}
\description{
Apply a recode rule to a raw-value vector.
}
