% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosswalk.R
\name{brfss_set_pool}
\alias{brfss_set_pool}
\title{Set the directory for raw BRFSS survey data}
\usage{
brfss_set_pool(path)
}
\arguments{
\item{path}{Character string pointing to the local data pool.}
}
\description{
Tells \code{brfssTools} where your local pool of raw survey data is stored.
This directory should contain the yearly survey files saved as \code{.csv}
objects (e.g., "BRFSS_2020.csv").
}
\examples{
\dontrun{
brfss_set_pool("Z:/Secure_Data/Oregon_Surveys/Harmonized_Pool")
}
}
