% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_state.R
\name{sampled_states}
\alias{sampled_states}
\title{States with a built-in sample shipped in the package}
\usage{
sampled_states()
}
\value{
Character vector of state postal codes.
}
\description{
States with a built-in sample shipped in the package
}
